Fix a scaling bug with the line numbers in text editors.

+++ source/dred_text_editor.h
@@ -43,4 +43,4 @@ bool dred_text_editor_find_and_replace_all(dred_text_editor* pTextEditor, const 
 
 
 // Sets the scale of the internal text.
-void dred_text_editor_set_text_scale(dred_text_editor* pTextEditor, double textScale);
+void dred_text_editor_set_text_scale(dred_text_editor* pTextEditor, float textScale);
